move auth related models to own folder
